add Migrations

+++ src/App/Transaction.Storage.Srv.App.Core.Aggregates.TransactionAggregate/Interfaces/IPositionBodyDto.cs
@@ -4,7 +4,7 @@ public interface IPositionBodyDto
 {
   public int? AccountId { get; }
 
-  public int Amount { get; }
+  public decimal Amount { get; }
 
   public int AssetId { get; }
 }
